fix: remove old table route
