--- src/routes/admin/users/+page.server.ts
@@ -1,1 +1,0 @@
-// retrieve user data here
