--- src/routes/admin/users/data.ts
@@ -1,230 +1,0 @@
-// Temporary dummy data
-import { type UserTable } from '$lib/dataTypes/EntityTypes';
-
-export const users: UserTable[] = [
-	{
-		lib_user_id: 1,
-		rfid: 100000001,
-		username: 'jdoe',
-		first_name: 'John',
-		middle_initial: 'A',
-		last_name: 'Doe',
-		phone_number: '1234567890',
-		is_enrolled: true,
-		is_active: true,
-		college: 'Engineering',
-		unit: 'Electrical Engineering',
-		program: 'BSEE',
-		user_type: 'Student'
-	},
-	{
-		lib_user_id: 2,
-		rfid: 100000002,
-		username: 'asmith',
-		first_name: 'Alice',
-		middle_initial: null,
-		last_name: 'Smith',
-		phone_number: '0987654321',
-		is_enrolled: true,
-		is_active: true,
-		college: 'Business',
-		unit: 'Accounting',
-		program: 'BSA',
-		user_type: 'Student'
-	},
-	{
-		lib_user_id: 3,
-		rfid: null,
-		username: 'tbrown',
-		first_name: 'Tom',
-		middle_initial: 'M',
-		last_name: 'Brown',
-		phone_number: '1122334455',
-		is_enrolled: false,
-		is_active: false,
-		college: 'Arts',
-		unit: 'Music',
-		program: 'BM',
-		user_type: 'Faculty'
-	},
-	{
-		lib_user_id: 4,
-		rfid: 100000003,
-		username: 'kmiller',
-		first_name: 'Karen',
-		middle_initial: null,
-		last_name: 'Miller',
-		phone_number: '6677889900',
-		is_enrolled: true,
-		is_active: true,
-		college: 'Sciences',
-		unit: 'Biology',
-		program: 'BSBio',
-		user_type: 'Student'
-	},
-	{
-		lib_user_id: 5,
-		rfid: 100000004,
-		username: 'jwilson',
-		first_name: 'James',
-		middle_initial: 'R',
-		last_name: 'Wilson',
-		phone_number: '3344556677',
-		is_enrolled: true,
-		is_active: false,
-		college: 'Engineering',
-		unit: 'Civil Engineering',
-		program: 'BSCE',
-		user_type: 'Alumni'
-	},
-	{
-		lib_user_id: 6,
-		rfid: null,
-		username: 'esanchez',
-		first_name: 'Emma',
-		middle_initial: 'L',
-		last_name: 'Sanchez',
-		phone_number: '4433221100',
-		is_enrolled: true,
-		is_active: true,
-		college: 'Medicine',
-		unit: 'Nursing',
-		program: 'BSN',
-		user_type: 'Faculty'
-	},
-	{
-		lib_user_id: 7,
-		rfid: 100000005,
-		username: 'gmartin',
-		first_name: 'George',
-		middle_initial: null,
-		last_name: 'Martin',
-		phone_number: '2211334455',
-		is_enrolled: true,
-		is_active: true,
-		college: 'Sciences',
-		unit: 'Physics',
-		program: null,
-		user_type: 'Staff'
-	},
-	{
-		lib_user_id: 8,
-		rfid: 100000006,
-		username: 'cperez',
-		first_name: 'Carla',
-		middle_initial: 'D',
-		last_name: 'Perez',
-		phone_number: '1234561111',
-		is_enrolled: true,
-		is_active: true,
-		college: 'Arts',
-		unit: 'Literature',
-		program: 'BA Literature',
-		user_type: 'Student'
-	},
-	{
-		lib_user_id: 9,
-		rfid: 100000007,
-		username: 'jdavis',
-		first_name: 'Jack',
-		middle_initial: null,
-		last_name: 'Davis',
-		phone_number: '9998887777',
-		is_enrolled: false,
-		is_active: false,
-		college: 'Engineering',
-		unit: 'Mechanical Engineering',
-		program: 'BSME',
-		user_type: 'Faculty'
-	},
-	{
-		lib_user_id: 10,
-		rfid: null,
-		username: 'mrobinson',
-		first_name: 'Mary',
-		middle_initial: 'E',
-		last_name: 'Robinson',
-		phone_number: '5544332211',
-		is_enrolled: true,
-		is_active: false,
-		college: 'Business',
-		unit: 'Finance',
-		program: 'BSFin',
-		user_type: 'Alumni'
-	},
-	{
-		lib_user_id: 11,
-		rfid: 100000008,
-		username: 'awhite',
-		first_name: 'Anne',
-		middle_initial: 'T',
-		last_name: 'White',
-		phone_number: '7776665555',
-		is_enrolled: true,
-		is_active: true,
-		college: 'Sciences',
-		unit: 'Chemistry',
-		program: 'BSChem',
-		user_type: 'Student'
-	},
-	{
-		lib_user_id: 12,
-		rfid: null,
-		username: 'hclark',
-		first_name: 'Harry',
-		middle_initial: 'P',
-		last_name: 'Clark',
-		phone_number: '6665554444',
-		is_enrolled: false,
-		is_active: false,
-		college: 'Arts',
-		unit: 'Theater',
-		program: 'BA Theater',
-		user_type: 'Staff'
-	},
-	{
-		lib_user_id: 13,
-		rfid: 100000009,
-		username: 'pmoore',
-		first_name: 'Paul',
-		middle_initial: null,
-		last_name: 'Moore',
-		phone_number: '4443332222',
-		is_enrolled: true,
-		is_active: true,
-		college: 'Engineering',
-		unit: 'Computer Engineering',
-		program: 'BSCpE',
-		user_type: 'Student'
-	},
-	{
-		lib_user_id: 14,
-		rfid: 100000010,
-		username: 'clopez',
-		first_name: 'Carmen',
-		middle_initial: 'N',
-		last_name: 'Lopez',
-		phone_number: '8889990000',
-		is_enrolled: true,
-		is_active: true,
-		college: 'Medicine',
-		unit: 'Pharmacy',
-		program: 'BSPharm',
-		user_type: 'Student'
-	},
-	{
-		lib_user_id: 15,
-		rfid: 100000011,
-		username: 'bjones',
-		first_name: 'Ben',
-		middle_initial: 'K',
-		last_name: 'Jones',
-		phone_number: '1231231234',
-		is_enrolled: false,
-		is_active: false,
-		college: 'Law',
-		unit: 'Legal Studies',
-		program: 'LLB',
-		user_type: 'Faculty'
-	}
-];
